+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Unreleased
 
+## Changed
+
+## Added
+- Windows: Added Windows-only keys. Use conditional compilation (`#[cfg(target_os = "windows")]`) to use them to not break the build on other OSs. F21-F24 keys and the XBUTTON1 & XBUTTON2 mouse buttons are now available
+## Fixed
+
 # 0.1.0
 We should have bumped the minor version with the last release. Sorry about that. Have a look at the changes of 0.0.15 if you come from an earlier version.